+++ android/src/main/kotlin/com/fluttercandies/photo_manager/permission/PermissionDelegate.kt
@@ -52,7 +52,7 @@ abstract class PermissionDelegate {
                 PackageManager.GET_PERMISSIONS
             )
         }
-        return packageInfo.requestedPermissions.contains(permission)
+        return packageInfo.requestedPermissions?.contains(permission) == true
     }
 
     /**