+++ ios/Classes/PhotoChangeObserver.h
@@ -1,6 +1,0 @@
-#import <Foundation/Foundation.h>
-#import "PMImport.h"
-
-@interface PhotoChangeObserver : NSObject
-- (void)initWithRegister:(NSObject <FlutterPluginRegistrar> *)registrar;
-@end